Add support for COWDict.
The primary problem is that the eval does not accept anything other
than a regular dict as a dict for globals. This I guess is for
performance reasons. Earlier we had the top level CalcDict evaluated
to a normal dict before being passed to eval. But this lead to cyclic
dependency issues, since the top level dict evaluation itself required
the top level dict for evaluation. In the current approach we pass the
top level dict as locals to the eval. The downside of this approach is
that when locals variables are created, they will pollute the YAML
document tree, overwrite top level elements and persist across
evaluations. To overcome this problem, during evaluation we create a
COWDict that wraps the YAML document tree, and stores local variables
and discards local them after evaluation.